changed all the models and view models and repository to use float for the price of the stocks and money. Also added a Time event for updating the stocks price every minute with a random number between 1.00 and 4000.00

+++ Models/FinancialPortfolioModel.cs
@@ -5,7 +5,7 @@ namespace StockTradingApplication.Models;
 public class FinancialPortfolioModel
 {
     public List<StockModel> Stocks { get; set; }
-    public decimal Money { get; set; }
+    public float Money { get; set; }
 
     public FinancialPortfolioModel()
     {
