+++ Repositories/StockModelRepository.cs
@@ -7,11 +7,11 @@ public class StockModelRepository : IRepository<StockModel, string>
 {
     private readonly List<StockModel> _stocks = new List<StockModel>
     {
-        new StockModel { Symbol = "AAPL", Quantity = 10, Price = 150.00m },
-        new StockModel { Symbol = "GOOGL", Quantity = 5, Price = 2800.00m },
-        new StockModel { Symbol = "MSFT", Quantity = 15, Price = 300.00m },
-        new StockModel { Symbol = "AMZN", Quantity = 20, Price = 3000.00m },
-        new StockModel { Symbol = "TSLA", Quantity = 5, Price = 500.00m }
+        new StockModel { Symbol = "AAPL", Quantity = 10, Price = 150.00f },
+        new StockModel { Symbol = "GOOGL", Quantity = 5, Price = 2800.00f },
+        new StockModel { Symbol = "MSFT", Quantity = 15, Price = 300.00f },
+        new StockModel { Symbol = "AMZN", Quantity = 20, Price = 3000.00f },
+        new StockModel { Symbol = "TSLA", Quantity = 5, Price = 500.00f }
     };
 
     public StockModel Get(string stockSymbol)
@@ -54,7 +54,7 @@ public class StockModelRepository : IRepository<StockModel, string>
             System.Windows.MessageBox.Show($"Stock {stockSymbol} does not exist in the repository.", "Error");
         }
     }
-    public IEnumerable<StockModel> GetAllBelowPrice(decimal stockPrice)
+    public IEnumerable<StockModel> GetAllBelowPrice(float stockPrice)
     {
         if (_stocks.Count == 0)
         {
@@ -63,7 +63,7 @@ public class StockModelRepository : IRepository<StockModel, string>
         }
         return _stocks.Where(x => x.Price < stockPrice);
     }
-    public IEnumerable<StockModel> GetAllAbovePrice(decimal stockPrice)
+    public IEnumerable<StockModel> GetAllAbovePrice(float stockPrice)
     {
         if (_stocks.Count == 0)
         {
